+++ lib/generators/espresso/vue3/component/templates/component_template.vue
@@ -1,0 +1,23 @@
+<template lang="pug">
+
+<%= css_class_name %>
+
+</template>
+
+<script>
+export default {
+  // components: {},
+  // mixins: [],
+  // props: {},
+  setup() {
+    return {}
+  },
+  // data() {
+  //   return {}
+  // },
+  // computed: {},
+  // created() {},
+  // mounted() {},
+  // methods: {}
+}
+</script>